requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.0
httpx==0.27.0
pytest==8.2.2
rich==13.7.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Scheduler
apscheduler==3.10.4

sse-starlette==2.1.0
redis==5.0.1
